+++ src/anyio/_backends/_asyncio.py
@@ -27,6 +27,7 @@ from types import TracebackType
 from typing import (
     IO,
     Any,
+    AsyncGenerator,
     Awaitable,
     Callable,
     Collection,
@@ -2076,9 +2077,11 @@ class TestRunner(abc.TestRunner):
         use_uvloop: bool = False,
         policy: Optional[asyncio.AbstractEventLoopPolicy] = None,
     ):
+        self._exceptions: List[BaseException] = []
         _maybe_set_event_loop_policy(policy, use_uvloop)
         self._loop = asyncio.new_event_loop()
         self._loop.set_debug(debug)
+        self._loop.set_exception_handler(self._exception_handler)
         asyncio.set_event_loop(self._loop)
 
     def _cancel_all_tasks(self) -> None:
@@ -2099,6 +2102,21 @@ class TestRunner(abc.TestRunner):
             if task.exception() is not None:
                 raise cast(BaseException, task.exception())
 
+    def _exception_handler(
+        self, loop: asyncio.AbstractEventLoop, context: dict[str, Any]
+    ) -> None:
+        if isinstance(context["exception"], Exception):
+            self._exceptions.append(context["exception"])
+
+    def _raise_async_exceptions(self) -> None:
+        # Re-raise any exceptions raised in asynchronous callbacks
+        if self._exceptions:
+            exceptions, self._exceptions = self._exceptions, []
+            if len(exceptions) == 1:
+                raise exceptions[0]
+            elif exceptions:
+                raise ExceptionGroup(exceptions)
+
     def close(self) -> None:
         try:
             self._cancel_all_tasks()
@@ -2107,27 +2125,55 @@ class TestRunner(abc.TestRunner):
             asyncio.set_event_loop(None)
             self._loop.close()
 
-    def call(
-        self, func: Callable[..., Awaitable[T_Retval]], *args: object, **kwargs: object
+    def run_asyncgen_fixture(
+        self,
+        fixture_func: Callable[..., AsyncGenerator[T_Retval, Any]],
+        kwargs: Dict[str, Any],
+    ) -> Iterable[T_Retval]:
+        async def fixture_runner() -> None:
+            agen = fixture_func(**kwargs)
+            try:
+                retval = await agen.asend(None)
+                self._raise_async_exceptions()
+            except BaseException as exc:
+                f.set_exception(exc)
+                return
+            else:
+                f.set_result(retval)
+
+            await event.wait()
+            try:
+                await agen.asend(None)
+            except StopAsyncIteration:
+                pass
+            else:
+                await agen.aclose()
+                raise RuntimeError("Async generator fixture did not stop")
+
+        f = self._loop.create_future()
+        event = asyncio.Event()
+        fixture_task = self._loop.create_task(fixture_runner())
+        self._loop.run_until_complete(f)
+        yield f.result()
+        event.set()
+        self._loop.run_until_complete(fixture_task)
+        self._raise_async_exceptions()
+
+    def run_fixture(
+        self,
+        fixture_func: Callable[..., Coroutine[Any, Any, T_Retval]],
+        kwargs: Dict[str, Any],
     ) -> T_Retval:
-        def exception_handler(
-            loop: asyncio.AbstractEventLoop, context: Dict[str, Any]
-        ) -> None:
-            exceptions.append(context["exception"])
-
-        exceptions: List[BaseException] = []
-        self._loop.set_exception_handler(exception_handler)
-        try:
-            retval: T_Retval = self._loop.run_until_complete(func(*args, **kwargs))
-        except Exception as exc:
-            retval = None  # type: ignore[assignment]
-            exceptions.append(exc)
-        finally:
-            self._loop.set_exception_handler(None)
-
-        if len(exceptions) == 1:
-            raise exceptions[0]
-        elif exceptions:
-            raise ExceptionGroup(exceptions)
-
+        retval = self._loop.run_until_complete(fixture_func(**kwargs))
+        self._raise_async_exceptions()
         return retval
+
+    def run_test(
+        self, test_func: Callable[..., Coroutine[Any, Any, Any]], kwargs: Dict[str, Any]
+    ) -> None:
+        try:
+            self._loop.run_until_complete(test_func(**kwargs))
+        except Exception as exc:
+            self._exceptions.append(exc)
+
+        self._raise_async_exceptions()